Revert "Fixed CA1812"
This reverts commit 6299072bb29e4f5a608b8deb006023e73d2e3f5c.
